changed api and config
