+++ src/api/group/index.js
@@ -1,6 +1,4 @@
 module.exports = {
   get: require('./get'),
-  post: require('./post'),
   delete: require('./delete'),
-  deleteGroup: require('./deleteGroup'),
 };
