+++ src/api/tag/index.js
@@ -1,5 +1,4 @@
 module.exports = {
   get: require('./get'),
-  post: require('./post'),
   delete: require('./delete'),
 };
