+++ src/api/user/index.js
@@ -5,4 +5,8 @@ module.exports = {
   update: require('./update'),
   updateSchema: require('./update-schema'),
   delete: require('./delete'),
+  addToGroup: require('./addToGroup'),
+  removeFromGroup: require('./removeFromGroup'),
+  addTag: require('./addTag'),
+  removeTag: require('./removeTag'),
 };
