+++ src/api/user/removeTag.js
@@ -1,10 +1,9 @@
 const queries = require('./queries');
 const userQueries = require('../user/queries');
 
-module.exports = async function addUserGroup(ctx) {
-  // TODO: add validation
+module.exports = async function deleteUserTag(ctx) {
   const userExternalId = ctx.params.user_id;
-  const groupName = ctx.params.group_name;
+  const tagName = ctx.params.tag_name.toLowerCase();
 
   const user = await userQueries.getUserByExternalId(userExternalId);
   if (!user) {
@@ -14,15 +13,15 @@ module.exports = async function addUserGroup(ctx) {
   }
 
   try {
-    await queries.addUserGroup(user.id, groupName);
+    await queries.deleteUserTag(user.id, tagName);
   } catch (error) {
     if (error.code === '23505') {
       ctx.response.status = 409;
-      ctx.fail({ group: 'user already in group' });
+      ctx.fail({ tag: 'user not in tag' });
       return;
     }
     throw error;
   }
 
-  ctx.success({ id: userExternalId, group: groupName });
+  ctx.success({ id: userExternalId, tag: tagName });
 };
