+++ src/api/webhook/index.js
@@ -4,4 +4,5 @@ module.exports = {
   get: require('./get'),
   delete: require('./delete'),
   update: require('./update'),
+  getOne: require('./getOne'),
 };
