+++ src/config/api.js
@@ -1,6 +1,6 @@
 const { getAssert } = require('../lib/util');
 
 module.exports = {
-  port: getAssert(process.env.API_PORT),
-  logLevel: getAssert(process.env.API_LOG_LEVEL),
+  port: 8080,
+  logLevel: 'error',
 };
