+++ src/config/db.js
@@ -1,9 +1,9 @@
 const { getAssert } = require('../lib/util');
 
 module.exports = {
-  host: getAssert(process.env.DB_HOST),
-  port: getAssert(process.env.DB_PORT),
-  database: getAssert(process.env.DB_NAME),
-  user: getAssert(process.env.DB_USER),
-  password: process.env.DB_PASSWORD,
+  host: 'localhost',
+  port: 5432,
+  database: 'notifications',
+  user: 'notificator',
+  password: null,
 };
