+++ src/config/notificator.js
@@ -1,10 +1,10 @@
 const { getAssert } = require('../lib/util');
 
 module.exports = {
-  maxWorkers: getAssert(process.env.NOTI_WORKERS),
-  getDataPollInterval: getAssert(process.env.NOTI_GET_DATA_POLL_INTERVAL),
-  getNextWorkerPollInterval: getAssert(process.env.NOTI_GET_NEXT_WORKER_POLL_INTERVAL),
-  workerDeadlockTimeout: getAssert(process.env.NOTI_WORKER_DEADLOCK_TIMEOUT),
-  workerPingInterval: getAssert(process.env.NOTI_WORKER_PING_INTERVAL),
-  logLevel: getAssert(process.env.NOTI_LOG_LEVEL),
+  maxWorkers: 2,
+  getDataPollInterval: 500,
+  getNextWorkerPollInterval: 100,
+  workerDeadlockTimeout: 60000,
+  workerPingInterval: 1000,
+  logLevel: 'error',
 };
