+++ src/lib/util.js
@@ -1,4 +1,3 @@
-//TODO: https://github.com/hiddentao/squel/issues/154 ?
 function pgArr(array, cast) {
   let arrStr =  !array ? '{}':JSON.stringify(array).replace('[', '{').replace(']', '}');
   if (cast) {
